+++ services/dedup/dedup.go
@@ -3,17 +3,14 @@
 package dedup
 
 import (
-	"fmt"
-	"sync"
-
-	"github.com/rudderlabs/rudder-go-kit/stats"
-
 	"github.com/rudderlabs/rudder-go-kit/config"
+	"github.com/rudderlabs/rudder-go-kit/stats"
 	"github.com/rudderlabs/rudder-server/services/dedup/badger"
+	"github.com/rudderlabs/rudder-server/services/dedup/scylla"
 	"github.com/rudderlabs/rudder-server/services/dedup/types"
 )
 
-type Mode string // skipcq: RVV-B0009
+type Mode string
 
 const (
 	Badger Mode = "Badger"
@@ -22,23 +19,21 @@ const (
 )
 
 // New creates a new deduplication service. The service needs to be closed after use.
-func New(conf *config.Config, stats stats.Stats) Dedup {
+func New(conf *config.Config, stats stats.Stats) (Dedup, error) {
 	mode := Mode(config.GetString("Dedup.Mode", string(Badger)))
 	switch mode {
 	case Badger:
-		return &dedup{
-			badgerDB: badger.NewBadgerDB(conf, stats, badger.DefaultPath()),
-			cache:    make(map[string]int64),
-		}
+		return badger.NewBadgerDB(conf, stats, badger.DefaultPath()), nil
 	case Scylla:
-		return nil
-	case Dual:
-		return nil
-	default:
-		return &dedup{
-			badgerDB: badger.NewBadgerDB(conf, stats, badger.DefaultPath()),
-			cache:    make(map[string]int64),
+		scylla, err := scylla.New(conf, stats)
+		if err != nil {
+			return nil, err
 		}
+		return scylla, nil
+	case Dual:
+		return nil, nil
+	default:
+		return badger.NewBadgerDB(conf, stats, badger.DefaultPath()), nil
 	}
 }
 
@@ -52,52 +47,4 @@ type Dedup interface {
 
 	// Close closes the deduplication service
 	Close()
-}
-
-type dedup struct {
-	badgerDB *badger.BadgerDB
-	cacheMu  sync.Mutex
-	cache    map[string]int64
-}
-
-func (d *dedup) Set(kv types.KeyValue) (bool, int64, error) {
-	d.cacheMu.Lock()
-	defer d.cacheMu.Unlock()
-	if previous, found := d.cache[kv.Key]; found {
-		return false, previous, nil
-	}
-	previous, found, err := d.badgerDB.Get(kv.Key)
-	if err != nil {
-		return false, 0, err
-	}
-	if !found {
-		d.cache[kv.Key] = kv.Value
-	}
-	return !found, previous, nil
-}
-
-func (d *dedup) Commit(keys []string) error {
-	d.cacheMu.Lock()
-	defer d.cacheMu.Unlock()
-
-	kvs := make([]types.KeyValue, len(keys))
-	for i, key := range keys {
-		value, ok := d.cache[key]
-		if !ok {
-			return fmt.Errorf("key %v has not been previously set", key)
-		}
-		kvs[i] = types.KeyValue{Key: key, Value: value}
-	}
-
-	err := d.badgerDB.Set(kvs)
-	if err == nil {
-		for _, kv := range kvs {
-			delete(d.cache, kv.Key)
-		}
-	}
-	return err
-}
-
-func (d *dedup) Close() {
-	d.badgerDB.Close()
 }
